eliminated most of the warnings.

+++ cass/sharedmemory_input.cpp
@@ -107,7 +107,7 @@ int cass::SharedMemoryInput::processDgram(Pds::Dgram* datagram)
   //read the datagram to the ringbuffer//
   Pds::Dgram& dg = *reinterpret_cast<Pds::Dgram*>(cassevent->datagrambuffer());
   memcpy(&dg,datagram,sizeof(Pds::Dgram));
-  if (datagram->xtc.sizeofPayload() > cass::DatagramBufferSize)
+  if (datagram->xtc.sizeofPayload() > static_cast<int>(cass::DatagramBufferSize))
     std::cout << "datagram size is bigger than the maximum buffer size of "<<cass::DatagramBufferSize/1024/1024<<" MB. Something is wrong"<<std::endl;
   memcpy(dg.xtc.payload(),datagram+1,datagram->xtc.sizeofPayload());
